+++ src/Phan/AST/UnionTypeVisitor.php
@@ -1175,7 +1175,7 @@ class UnionTypeVisitor extends AnalysisVisitor
             // And use those closures to infer the (possibly transformed) types
             $template_type_list = [];
             foreach ($template_type_resolvers as $template_type_resolver) {
-                $template_type_list[] = $template_type_resolver($arg_type_list);
+                $template_type_list[] = $template_type_resolver($arg_type_list, $this->context);
             }
 
             // Create a new type that assigns concrete